Modified targeting behavior to increase optimization of salvaging and mining routes (Nearest Neighbor)
It's not the most optimal solution but it is significantly more effective than the "closest to ship" solution.  Technically this implements the Nearest Neighbor algorithm as applied to the travelling salesman problem.  Which is fancy talk for it selects the asteroid or wreck closest to the one it just finished, rather than zig zagging back and forth across the sector wasting time instead of mining/salvaging.

Actual in gae increase in harvesting rates have not been calculated, but I went from taking over an hour to mine a system to about 15 minutes.